ENG-14773: Push createUpdatePartitionCountTasksForSites down to SnapshotWritePlan
